Added Player movement, fixed some settings
