+++ Assets/InputActions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a722e5d5a6cff8b4d9224d099af47a28
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
