+++ Assets/InputActions/Default.inputactions.meta
@@ -1,0 +1,14 @@
+fileFormatVersion: 2
+guid: a9fa31d722d3dc845ba96e17c1edcadd
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
+  generateWrapperCode: 0
+  wrapperCodePath: 
+  wrapperClassName: 
+  wrapperCodeNamespace:
