--- Assets/Scripts/Player/PlayerStates.cs
@@ -1,40 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using UnityEngine;
-
-public enum PlayerFacing
-{
-	UP,
-	LEFT,
-	RIGHT,
-	DOWN
-}
-
-public enum PlayerState
-{
-	WALK,
-	IDLE,
-	DIE,
-	OTHER
-}
-
-public class Player : MonoBehaviour
-{
-	public PlayerFacing direction = PlayerFacing.DOWN;
-	public PlayerState state = PlayerState.IDLE;
-
-	// Start is called before the first frame update
-	void Start()
-	{
-
-	}
-
-	// Update is called once per frame
-	void Update()
-	{
-
-	}
-}
