+++ Logs/Packages-Update.log
@@ -48,3 +48,17 @@ The following packages were updated:
   com.unity.2d.animation from version 5.0.1 to 5.0.3
   com.unity.2d.psdimporter from version 4.0.1 to 4.0.2
   com.unity.2d.spriteshape from version 5.0.1 to 5.1.0
+
+=== Fri Jul  9 17:33:33 2021
+
+Packages were changed.
+Update Mode: updateDependencies
+
+The following packages were updated:
+  com.unity.2d.animation from version 5.0.3 to 5.0.5
+  com.unity.2d.spriteshape from version 5.1.0 to 5.1.2
+  com.unity.collab-proxy from version 1.3.9 to 1.5.7
+  com.unity.ide.visualstudio from version 2.0.5 to 2.0.9
+  com.unity.test-framework from version 1.1.19 to 1.1.26
+  com.unity.textmeshpro from version 3.0.1 to 3.0.6
+  com.unity.timeline from version 1.4.4 to 1.4.8
