+++ Packages/manifest.json
@@ -1,18 +1,19 @@
 {
   "dependencies": {
-    "com.unity.2d.animation": "5.0.3",
+    "com.unity.2d.animation": "5.0.5",
     "com.unity.2d.pixel-perfect": "4.0.1",
     "com.unity.2d.psdimporter": "4.0.2",
     "com.unity.2d.sprite": "1.0.0",
-    "com.unity.2d.spriteshape": "5.1.0",
+    "com.unity.2d.spriteshape": "5.1.2",
     "com.unity.2d.tilemap": "1.0.0",
-    "com.unity.collab-proxy": "1.3.9",
+    "com.unity.collab-proxy": "1.5.7",
     "com.unity.ide.rider": "2.0.7",
-    "com.unity.ide.visualstudio": "2.0.5",
+    "com.unity.ide.visualstudio": "2.0.9",
     "com.unity.ide.vscode": "1.2.3",
-    "com.unity.test-framework": "1.1.19",
-    "com.unity.textmeshpro": "3.0.1",
-    "com.unity.timeline": "1.4.4",
+    "com.unity.inputsystem": "1.0.2",
+    "com.unity.test-framework": "1.1.26",
+    "com.unity.textmeshpro": "3.0.6",
+    "com.unity.timeline": "1.4.8",
     "com.unity.ugui": "1.0.0",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.androidjni": "1.0.0",
